% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/QPGlobal.R
\name{QPGlobal}
\alias{QPGlobal}
\title{Finding the global quasi-potential}
\usage{
QPGlobal(local.surfaces, unstable.eq.x, unstable.eq.y, x.bound, y.bound)
}
\arguments{
\item{local.surfaces}{A list of local quasi-potential surfaces, each of which is stored in discretized form as a matrix.}

\item{unstable.eq.x}{A vector of the x-coordinates of the unstable equilibria.  Must be in the same order as unstable.eq.y.}

\item{unstable.eq.y}{A vector of the y-coordinates of the unstable equilibria.  Must be in the same order as unstable.eq.x.}

\item{x.bound}{A two-element vector with the minimum and maximum x values used for computing the quasi-potential.}

\item{y.bound}{A two-element vector with the minimum and maximum y values used for computing the quasi-potential.}
}
\description{
This function allows you to find the global quasi-potential values for several local quasi-potential surfaces
}
\examples{
#QPGlobal(list(local.1,local.2),c(0,4),c(0,4),c(-1,5),c(-1,5))
}
\keyword{#}
\keyword{Global}
\keyword{quasi-potential}

