% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/QPInterp.R
\name{QPInterp}
\alias{QPInterp}
\title{Quasi-potential interpolation}
\usage{
QPInterp(X, Y, x.bound, y.bound, surface)
}
\arguments{
\item{X}{the value of the x to interpolate}

\item{Y}{the value of the y to interpolate}

\item{x.bound}{two-element vector with respective minimum and maximum x values.}

\item{y.bound}{two-element vector with respective minimum and maximum y values.}

\item{surface}{the surface to interpolated, from \code{\link{QPGlobal}}}
}
\description{
This function estimates the quasi-potential value for any x- and y-values
}
\details{
this function uses simple bilinear interpolation for estimattion of any x- and y-value.
}
\keyword{interpolation}

