% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QPot-package.R
\docType{package}
\name{QPot-package}
\alias{QPot}
\alias{QPot-package}
\title{QPot: Quasi-Potential Analysis for Stochastic Differential Equations}
\description{
Tools to 1) simulate and visualize stochastic differential equations and 2) determine stability of equilibria using the ordered-upwind method to compute the quasi-potential.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.r-project.org}
  \item \url{https://github.com/bmarkslash7/QPot}
  \item Report bugs at \url{https://github.com/bmarkslash7/QPot/issues}
}

}
\author{
\strong{Maintainer}: Christopher Stieha \email{stieha@hotmail.com}

Authors:
\itemize{
  \item Christopher Moore \email{life.dispersing@gmail.com}
  \item Ben Nolting
  \item Maria Cameron
  \item Karen Abbott
}

Other contributors:
\itemize{
  \item James Gregson (author of expression_parser library: https://github.com/jamesgregson/expression_parser) [copyright holder]
}

}
\keyword{internal}
