% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TSDensity.R
\name{TSDensity}
\alias{TSDensity}
\title{Density plot from simulation of two-dimensional stochastic differential equations}
\usage{
TSDensity(mat, dim = 1, contour.levels = 15, col2d = c("blue", "yellow",
  "orange", "red"), contour.lwd = 0.5, contour.lines = TRUE)
}
\arguments{
\item{mat}{a matrix output from \code{\link{TSTraj}}.}

\item{dim}{dimensions of the plot; \code{dim=1} plots simple density histogram or \code{dim=2} plots the density in state space (i.e., \code{X} and \code{Y} respectively on the abscissa and ordinate axes).}

\item{contour.levels}{the number of contour levels for the two-dimensional plots (i.e., when \code{dim=2}).}

\item{col2d}{vector of colors to be used in the plot.}

\item{contour.lwd}{line width of contour lines if \code{contour.lines=TRUE}.}

\item{contour.lines}{if TRUE, then black countour lines added to the graph.}
}
\description{
This function allows you to create density plots for the simulation of two-dimensional stochastic differential equations in \code{\link{TSTraj}}
}
\examples{
# First, the parameter values
state <- c(x = 3 , y = 3)
model.parms <- c(mu = 0.2)
model.sigma <- 0.1
model.deltat <- 0.005

# Second, create the model in the sense of deSolve
LVModel <- function(t, state, parms) {
with(as.list(c(state, parms)), {
dx <- -(y-5) + mu*(x-4)*(1-((x-4)^2)-((y-5)^2))
dy <- (x-4) + mu*(y-5)*(1-((x-4)^2)-((y-5)^2))
list(c(dx,dy))
})
}
}
\keyword{Density}
\keyword{of}
\keyword{plot}
\keyword{simulations}
\keyword{stochastic}

