% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TSPlot.R
\name{TSPlot}
\alias{TSPlot}
\title{Plot simulation of two-dimensional stochastic differential equations}
\usage{
TSPlot(mat, deltat, dim = 1, y.lim = NA, x.lab = "time", dens = TRUE,
  lwd = 2, line.alpha = 130, zero.axes = TRUE)
}
\arguments{
\item{mat}{a matrix output from \code{\link{TSTraj}}.}

\item{deltat}{numeric value indicating the frequency of stochastic perturbation, as \eqn{\Delta t}, used in the function to recaluculate axes if applicable.}

\item{dim}{dimensions of the plot; \code{dim=1} to plot a timeseries with \code{X} and \code{Y} on the ordinate axis or \code{dim=2} to plot the trjectories in state space (i.e., \code{X} and \code{Y} respectively on the abscissa and ordinate axes).}

\item{y.lim}{for \code{dim=1}, allows user to specify the range of the y-axis as a two-element vector.}

\item{x.lab}{for \code{dim=1}, allows user to specify the axis as "time" or "steps," with steps being \eqn{time \times \Delta t}}

\item{dens}{if \code{dens=TRUE}, plots a horizontal one-dimensional density plot adjacent to the timerseries.}

\item{lwd}{line width.}

\item{line.alpha}{transparency of lines from 0--255.}

\item{zero.axes}{if TRUE, then axes plotted at \code{X=0} and \code{Y=0}.}
}
\description{
This function allows you to plot the simulation of simulate two-dimensional stochastic differential equations from \code{\link{TSTraj}}
}
\examples{
# First, the parameter values
state <- c(x = 3 , y = 3)
model.parms <- c(mu = 0.2)
model.sigma <- 0.1
model.deltat <- 0.005
}
\keyword{plot}
\keyword{simulations}
\keyword{stochastic}

