% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VecDecomAll.R
\name{VecDecomAll}
\alias{VecDecomAll}
\title{Vector decomposition and remainder fields}
\usage{
VecDecomAll(surface, x.rhs, y.rhs, x.bound, y.bound)
}
\arguments{
\item{surface}{matrix output from QPGlobal.}

\item{x.rhs}{COPY FROM QPOTENTIAL().}

\item{y.rhs}{COPY FROM QPOTENTIAL().}

\item{x.bound}{two-element vector with respective minimum and maximum x values.}

\item{y.bound}{two-element vector with respective minimum and maximum y values.}
}
\description{
This function calculates the vector, gradient, and remainder fields.
}
\examples{
#Example 1 from article
equationx = "1.5*x*(1.0-(x/45.0))-(y*x*5.0)/(18.0+x)"
equationy = "-4.0*y+((10.0*x*y)/(18.0+x))"
}
\keyword{field}
\keyword{field,}
\keyword{remainder}
\keyword{vector}

