% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VecDecomGrad.R
\name{VecDecomGrad}
\alias{VecDecomGrad}
\title{Vector decomposition and remainder fields}
\usage{
VecDecomGrad(surface)
}
\arguments{
\item{surface}{matrix output from QPGlobal.}
}
\description{
This function calculates the gradient field.
}
\examples{
#Example 1 from article. Equations are not required
equationx = "1.5*x*(1.0-(x/45.0))-(y*x*5.0)/(18.0+x)"
equationy = "-4.0*y+((10.0*x*y)/(18.0+x))"
#e1.global <- matrix()
#VDG <- VecDecomGrad(e1.global)
#VecDecomPlot(field=list(VDG[,,1],VDG[,,2]), dens=c(50,50), x.bound=c(-0.5,20), y.bound=c(-0.5,20))
}
\keyword{decomposition,}
\keyword{field}
\keyword{gradient}
\keyword{vector}

