% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VecDecompAll.R
\name{VecDecompAll}
\alias{VecDecompAll}
\title{Vector decomposition and remainder fields}
\usage{
VecDecompAll(surface, x.rhs, y.rhs, x.bound, y.bound)
}
\arguments{
\item{surface}{matrix output from QPGlobal.}

\item{x.rhs}{COPY FROM QPOTENTIAL().}

\item{y.rhs}{COPY FROM QPOTENTIAL().}

\item{x.bound}{two-element vector with respective minimum and maximum x values.}

\item{y.bound}{two-element vector with respective minimum and maximum y values.}
}
\description{
This function calculates the vector, gradient, and remainder fields.
}
\examples{
#Example 1 from article
equationx = "1.5*x*(1.0-(x/45.0))-(y*x*5.0)/(18.0+x)"
equationy = "-4.0*y+((10.0*x*y)/(18.0+x))"
#e1.global <- matrix()
#VDAll <- VecDecompAll(surface=e1.global, x.rhs=testequationx, y.rhs=testequationy ,x.bound=c(-0.5,20), y.bound=c(-0.5,20))
## 0.6.1 vector field. See VecDecompVec()
#VecDecompPlot(field=list(VDV[,,1],VDV[,,2]), dens=c(50,50), x.bound=c(-0.5,20), y.bound=c(-0.5,20))
## 0.6.2 gradient field.  See VecDecompGrad()
#VecDecompPlot(field=list(VDV[,,3],VDV[,,4]), dens=c(50,50), x.bound=c(-0.5,20), y.bound=c(-0.5,20))
## 0.6.3 remainder field.  See VecDecompRem()
#VecDecompPlot(field=list(VDV[,,5],VDV[,,6]), dens=c(50,50), x.bound=c(-0.5,20), y.bound=c(-0.5,20))
}
\keyword{field}
\keyword{field,}
\keyword{remainder}
\keyword{vector}

