% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VecDecompGrad.R
\name{VecDecompGrad}
\alias{VecDecompGrad}
\title{Vector decomposition and remainder fields}
\usage{
VecDecompGrad(surface)
}
\arguments{
\item{surface}{matrix output from QPGlobal.}
}
\description{
This function calculates the gradient field.
}
\keyword{decompoosition,}
\keyword{field}
\keyword{gradient}
\keyword{vector}

