% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VecDecompPlot.R
\name{VecDecompPlot}
\alias{VecDecompPlot}
\title{Plotting function for vector decomposition and remainder fields}
\usage{
VecDecompPlot(field, dens, x.bound, y.bound, x.lim, y.lim,
  arrow.type = "equal", tail.length = 1, ...)
}
\arguments{
\item{field}{list output from \code{\link{VecDecompAll}}.}

\item{dens}{two-element vector respectively specifying the number of respective arrows in the x and y directions.}

\item{x.bound}{two-element vector for the x domain boundries used for the quasi-potential simulation.}

\item{y.bound}{two-element vector for the y domain boundries used for the quasi-potential simulation.}

\item{x.lim}{I DO NOT KNOW BUT NOW devtools::check() IS OK}

\item{y.lim}{I DO NOT KNOW BUT NOW devtools::check() IS OK}

\item{arrow.type}{sets the type of line segments plotted. If set to "proportional" the length of the line segments reflects the magnitude of the derivative. If set to "equal" the line segments take equal lengths, simply reflecting the gradient of the derivative(s). Defaults to "equal".}

\item{tail.length}{multiplies the current length of the tail (both proportional and equal arrow.types) by the specified factor.  The argument defaults to 1, which is length of the longest vector within the domain boundaries (i.e., the entire field).}

\item{...}{passes arguments to both \code{\link{plot}} and \code{\link{arrows}}.}
}
\description{
This function calculates the vector and remainder fields.
}
\examples{
x.limits <- c(-10,10)
y.limits <- c(-10,10)
length.xy <- c(15,15)
x <- seq(x.limits[1], x.limits[2], length.out = length.xy[1]) + 0.5
y <- seq(y.limits[1], y.limits[2], length.out = length.xy[2]) + 0.5
eqns <- list("(y^2)/(x)" , "(x^2)/(y)")
f <- function(x, y) { x <- (y^2)/(x) ; y <- (x^2)/(y) }
z <- outer(x, y, f)
#VecDecomp(z)
#VecDecomp(z,eqns,mesh.xy,x.limits,y.limits)
}
\keyword{field}
\keyword{plot}
\keyword{plot,}
\keyword{remainder}
\keyword{vector}

