% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VecDecompVec.R
\name{VecDecompVec}
\alias{VecDecompVec}
\title{Vector decomposition and remainder fields}
\usage{
VecDecompVec(x.num.steps, y.num.steps, x.rhs, y.rhs, x.bound, y.bound)
}
\arguments{
\item{x.num.steps}{COPY FROM QPOTENTIAL().}

\item{y.num.steps}{COPY FROM QPOTENTIAL().}

\item{x.rhs}{COPY FROM QPOTENTIAL().}

\item{y.rhs}{COPY FROM QPOTENTIAL().}

\item{x.bound}{two-element vector with respective minimum and maximum x values.}

\item{y.bound}{two-element vector with respective minimum and maximum y values.}
}
\description{
This function calculates the vector field.
}
\keyword{decompoosition,}
\keyword{field}
\keyword{vector}

