% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Model2String.R
\name{Model2String}
\alias{Model2String}
\title{Function that converts differential equations from
function-format to string-format
Specifically, it reads in a function, searches for the differential
equations within the function, and returns a list of strings containing
the differential equations from the function.
Will also replace parameters of those equations with numerical values.}
\usage{
Model2String(model.function, parms = "NULL", x.lhs.term = "dx",
  y.lhs.term = "dy", supress.print = FALSE)
}
\arguments{
\item{model.function}{function containing the differential equations as given to TSTraj()}

\item{parms}{a named vector of paramters and their respective values for the deterministic equations.}

\item{x.lhs.term}{string containing the left hand side of the first equation to search for, default is 'dx'}

\item{y.lhs.term}{string containing the left hand side of the second equation to search for, default is 'dx'}

\item{supress.print}{Default it FALSE, supress output.  TRUE prints out equations from function}
}
\value{
equations a list with two elements, the first is the x equation, the second is the y equation
}
\description{
Function that converts differential equations from
function-format to string-format
Specifically, it reads in a function, searches for the differential
equations within the function, and returns a list of strings containing
the differential equations from the function.
Will also replace parameters of those equations with numerical values.
}

