% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TSTraj.R
\name{TSTraj}
\alias{TSTraj}
\title{Simulate two-dimensional stochastic differential equations}
\usage{
TSTraj(y0, time, deltat, x.rhs, y.rhs, parms, sigma, lower.bound = NA,
  upper.bound = NA)
}
\arguments{
\item{y0}{the a two-element vector of the initial conditions for the state variables.}

\item{time}{numeric value indicating the total time over which the simulation is to be run.}

\item{deltat}{numeric value indicating the frequency of stochastic perturbation, as \eqn{\Delta t}.}

\item{x.rhs}{A string containing the right hand side of the equation for x.}

\item{y.rhs}{A string containing the right hand side of the y equation.}

\item{parms}{a named vector of paramters and their respective values for the deterministic equations.}

\item{sigma}{numberic value specifying the noise intensity.}

\item{lower.bound}{numeric value specifying a lower bound in the simulation.}

\item{upper.bound}{numeric value specifying an upper bound in the simulation.}
}
\description{
This function allows you to simulate two-dimensional stochastic differential equations.
}
\examples{
# First, the parameter values
state <- c(x = 3 , y = 3)
model.parms <- c(mu = 0.2)
model.sigma <- 0.1
model.deltat <- 0.005
}
\keyword{Stochastic}
\keyword{simulation}

