% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TSTraj.R
\name{TSTraj}
\alias{TSTraj}
\title{Simulate two-dimensional stochastic differential equations}
\usage{
TSTraj(y0, time, deltat, x.rhs, y.rhs, parms = NA, sigma, lower.bound = NA,
  upper.bound = NA)
}
\arguments{
\item{y0}{the a two-element vector of the initial conditions for the state variables.}

\item{time}{numeric value indicating the total time over which the simulation is to be run.}

\item{deltat}{numeric value indicating the frequency of stochastic perturbation, as \eqn{\Delta t}.}

\item{x.rhs}{A string containing the right hand side of the equation for x.}

\item{y.rhs}{A string containing the right hand side of the y equation.}

\item{parms}{a named vector of paramters and their respective values for the deterministic equations.}

\item{sigma}{numberic value specifying the noise intensity.}

\item{lower.bound}{numeric value specifying a lower bound in the simulation.}

\item{upper.bound}{numeric value specifying an upper bound in the simulation.}
}
\description{
This function allows you to simulate two-dimensional stochastic differential equations.
}
\examples{
# First, the parameter values
state <- c(x = 3 , y = 3)
model.sigma <- 0.1
model.deltat <- 0.005

# Second, write out the deterministic skeleton of the equations to be simulated
#Example 1 from article
equationx = "1.5*x*(1.0-(x/45.0))-(y*x*5.0)/(18.0+x)"
equationy = "-4.0*y+((10.0*x*y)/(18.0+x))"

# Third, Run it
LVModelOut <- TSTraj(y0=state, time=250, deltat=model.deltat,
x.rhs=equationx, y.rhs=equationy, sigma=model.sigma)

#Can also input x.rhs and y.rhs as strings that contain parameter names
#and include parms with names and values of parameters
model.state <- c(x=1 , y=2)
model.parms <- c(alpha=1.54, beta=10.14, delta=1, kappa=1, gamma=0.476, mu=0.112509)
model.sigma <- 0.05
model.time <- 1000
model.deltat <- 0.2

test.eqn.x = "(alpha*x)*(1-(x/beta)) - ((delta*(x^2)*y)/(kappa + (x^2)))"
test.eqn.y = "((gamma*(x^2)*y)/(kappa + (x^2))) - mu*(y^2)"

ts.out.ex1 <- TSTraj(y0= model.state, time=model.time, deltat=model.deltat, x.rhs=test.eqn.x, y.rhs= test.eqn.y, parms=model.parms, sigma=model.sigma)
}
\keyword{Stochastic}
\keyword{simulation}

