% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VecDecompAll.R
\name{VecDecompAll}
\alias{VecDecompAll}
\title{Vector decomposition and remainder fields}
\usage{
VecDecompAll(surface, x.rhs, y.rhs, x.bound, y.bound)
}
\arguments{
\item{surface}{matrix output from QPGlobal.}

\item{x.rhs}{COPY FROM QPOTENTIAL().}

\item{y.rhs}{COPY FROM QPOTENTIAL().}

\item{x.bound}{two-element vector with respective minimum and maximum x values.}

\item{y.bound}{two-element vector with respective minimum and maximum y values.}
}
\description{
This function calculates the vector, gradient, and remainder fields.
}
\keyword{field}
\keyword{field,}
\keyword{remainder}
\keyword{vector}

