% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VecDecompVec.R
\name{VecDecompVec}
\alias{VecDecompVec}
\title{Vector decomposition and remainder fields}
\usage{
VecDecompVec(x.num.steps, y.num.steps, x.rhs, y.rhs, x.bound, y.bound)
}
\arguments{
\item{x.num.steps}{COPY FROM QPOTENTIAL().}

\item{y.num.steps}{COPY FROM QPOTENTIAL().}

\item{x.rhs}{COPY FROM QPOTENTIAL().}

\item{y.rhs}{COPY FROM QPOTENTIAL().}

\item{x.bound}{two-element vector with respective minimum and maximum x values.}

\item{y.bound}{two-element vector with respective minimum and maximum y values.}
}
\description{
This function calculates the vector field.
}
\examples{
#Example 1 from article
equationx = "1.5*x*(1.0-(x/45.0))-(y*x*5.0)/(18.0+x)"
equationy = "-4.0*y+((10.0*x*y)/(18.0+x))"
VDV <- VecDecompVec(x.num.steps=4100, y.num.steps=4100, x.rhs=equationx,
 y.rhs=equationy, x.bound=c(-0.5,20), y.bound=c(-0.5,20))
VecDecompPlot(field=list(VDV[,,1],VDV[,,2]), dens=c(50,50),
 x.bound=c(-0.5,20), y.bound=c(-0.5,20))
}
\keyword{decompoosition,}
\keyword{field}
\keyword{vector}

