% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VecDecomVec.R
\name{VecDecomVec}
\alias{VecDecomVec}
\title{Vector decomposition and remainder fields}
\usage{
VecDecomVec(x.num.steps, y.num.steps, x.rhs, y.rhs, x.bound, y.bound)
}
\arguments{
\item{x.num.steps}{The number of steps between the minimum and maximum x value defined in x range.}

\item{y.num.steps}{The number of steps between the minimum and maximum y value defined in y range.}

\item{x.rhs}{A string containing the right hand side of the equation for x.}

\item{y.rhs}{A string containing the right hand side of the y equation.}

\item{x.bound}{two-element vector with respective minimum and maximum x values.}

\item{y.bound}{two-element vector with respective minimum and maximum y values.}
}
\description{
This function calculates the vector field.
}
\examples{
#Example 1 from article
equationx = "1.5*x*(1.0-(x/45.0))-(y*x*5.0)/(18.0+x)"
equationy = "-4.0*y+((10.0*x*y)/(18.0+x))"
VDV <- VecDecomVec(x.num.steps=4100, y.num.steps=4100, x.rhs=equationx,
 y.rhs=equationy, x.bound=c(-0.5,20), y.bound=c(-0.5,20))
VecDecomPlot(field=list(VDV[,,1],VDV[,,2]), dens=c(50,50),
 x.bound=c(-0.5,20), y.bound=c(-0.5,20))
}
\keyword{decompoosition,}
\keyword{field}
\keyword{vector}

